Improve xdlops_gemm lowering logic take offset in matrix C into consideration.
